Changed average review to be a double type
This gives a more accurate average rating. The front end can take this number and make it as many decimal points as they want

+++ backend/backend/src/main/java/coms309/trailtraveler/backend/controller/TrailController.java
@@ -56,7 +56,7 @@ public class TrailController {
 	}
 	
 	@GetMapping("trail/averageRating/{trailName}")
-	int getAverageTrailRating(@PathVariable String trailName) {
+	double getAverageTrailRating(@PathVariable String trailName) {
 		return ts.retrieveTrailRating(trailName);
 	}
 }
